requirements.txt
appdirs==1.4.4
argon2-cffi==19.2.0
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.13.2
chardet==4.0.0
cryptography==3.1.1
distlib==0.3.1
dj-database-url==0.5.0
Django==2.2.5
django-ckeditor==6.0.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-request==1.5.6
django-shopping-cart==0.1
djangorestframework==3.12.1
Faker==3.0.0
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
Pillow==6.2.1
pipenv==2020.8.13
psycopg2==2.8.6
pycparser==2.19
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
requests==2.25.1
six==1.13.0
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.26.2
virtualenv==20.0.31
virtualenv-clone==0.5.4
whitenoise==5.2.0
wincertstore==0.2
